move to normal net income tree
