--- policyengine_us/reforms/states/il/__init__.py
@@ -1,3 +1,0 @@
-from .il_child_tax_credit import (
-    create_il_child_tax_credit_reform,
-)
